Major change to upstream module approach.
Now using git submodules to keep up-to-date with latest otel code in a relatively convenient way.
Some dust is still settling, but the new approach to building is documented in the CONTRIBUTING.md.
Next up: fixing the circleci build, documenting more about this new stuff.

Side note: the upstream work I contributed to otel-js-contrib around removeEventListener is present here now
(before it's even been released), so in order for the tests to pass, this commit includes adjustments to the local main.js
to remove the hacky patch I had for that issue in favor of the peer-reviewed and stronger code in upstream.